requirements.txt
asgiref==3.8.1
Django==5.0.7
django-tinymce==4.1.0
jmespath==1.0.1
pillow==10.4.0
gunicorn==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.2
